AssetElement to render assets list and avoid code duplication
