+++ app/components/AssetElement/index.js
@@ -1,0 +1,74 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { TouchableOpacity, StyleSheet, View } from 'react-native';
+import { colors } from '../../styles/common';
+import Icon from 'react-native-vector-icons/Ionicons';
+
+const styles = StyleSheet.create({
+	itemWrapper: {
+		flex: 1,
+		flexDirection: 'row',
+		paddingHorizontal: 15,
+		paddingVertical: 10,
+		borderBottomWidth: StyleSheet.hairlineWidth,
+		borderBottomColor: colors.borderColor
+	},
+	arrow: {
+		flex: 1,
+		alignSelf: 'flex-end'
+	},
+	arrowIcon: {
+		marginTop: 16
+	}
+});
+
+/**
+ * Customizable view to render assets in lists
+ */
+export default class AssetElement extends Component {
+	static propTypes = {
+		/**
+		 * Content to display in the list element
+		 */
+		children: PropTypes.node,
+		/**
+		 * Object being rendered
+		 */
+		asset: PropTypes.object,
+		/**
+		 * Callback triggered on long press
+		 */
+		onPress: PropTypes.func,
+		/**
+		 * Callback triggered on long press
+		 */
+		onLongPress: PropTypes.func
+	};
+
+	handleOnPress = () => {
+		const { onPress, asset } = this.props;
+		onPress && onPress(asset);
+	};
+
+	handleOnLongPress = () => {
+		const { onLongPress, asset } = this.props;
+		onLongPress && onLongPress(asset);
+	};
+
+	render = () => {
+		const { children, asset } = this.props;
+		return (
+			<TouchableOpacity
+				onPress={this.handleOnPress}
+				onLongPress={this.handleOnLongPress}
+				style={styles.itemWrapper}
+				key={`asset=${asset.address}`}
+			>
+				{children}
+				<View styles={styles.arrow}>
+					<Icon name="ios-arrow-forward" size={24} color={colors.fontTertiary} style={styles.arrowIcon} />
+				</View>
+			</TouchableOpacity>
+		);
+	};
+}
